refactor: :building_construction: improve code structure
