+++ src/events/ready.ts
@@ -1,0 +1,8 @@
+import { Client } from 'discord.js';
+import logger from '../utils/logger';
+
+const onReady = (client: Client) => {
+  logger.info(`Logged in as ${client.user?.tag}!`);
+};
+
+export default onReady;
